syncing local and server

+++ tracker/network_tracker.c
@@ -903,7 +903,7 @@ void check_add_peer_q(_TNT_t * tnt) {
 	tracker_data_t * queue_item ;
 	while ( (queue_item = asyncqueue_pop(q)) != NULL) {
 
-		printf("NETWORK -- sending add peer (new peer %d) to client %d\n", queue_item->client_id, (int)queue_item->data);
+		printf("NETWORK -- sending add peer (new peer %d) to client %d\n", queue_item->client_id, (int)(long)queue_item->data);
 		peer_t * client = get_peer_by_id(tnt->peer_table, queue_item->client_id);
 		if (!client) {
 			fprintf(stderr, "check_add_peer_q: failed to find destination client\n");
@@ -920,7 +920,7 @@ void check_remove_peer_q(_TNT_t * tnt) {
 	tracker_data_t * queue_item ;
 	while ( (queue_item = asyncqueue_pop(q)) != NULL) {
 
-		printf("NETWORK -- sending remove peer (delete peer %d) to client %d\n", queue_item->client_id, (int)queue_item->data);
+		printf("NETWORK -- sending remove peer (delete peer %d) to client %d\n", queue_item->client_id, (int)(long)queue_item->data);
 		peer_t * client = get_peer_by_id(tnt->peer_table, queue_item->client_id);
 		if (!client) {
 			fprintf(stderr, "check_add_peer_q: failed to find destination client\n");
